+++ tests/PSRule.Tests/ObjectHelperTests.cs
@@ -1,7 +1,4 @@
 ﻿using PSRule.Commands;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using Xunit;
 
 namespace PSRule